@@ -1,88 +1,99 @@
 {
-	"name": "@discordjs/builders",
-	"version": "0.12.0",
-	"description": "A set of builders that you can use when creating your bot",
-	"scripts": {
-		"build": "tsup",
-		"test": "jest --pass-with-no-tests",
-		"lint": "prettier --check . && eslint src __tests__ --ext mjs,js,ts",
-		"format": "prettier --write . && eslint src __tests__ --ext mjs,js,ts --fix",
-		"docs": "typedoc --json docs/typedoc-out.json src/index.ts && node scripts/docs.mjs",
-		"prepublishOnly": "yarn build && yarn lint && yarn test",
-		"changelog": "git cliff --prepend ./CHANGELOG.md -u -c ./cliff.toml -r ../../ --include-path 'packages/builders/*'"
-	},
-	"main": "./dist/index.js",
-	"module": "./dist/index.mjs",
-	"types": "./dist/index.d.ts",
-	"exports": {
-		"import": "./dist/index.mjs",
-		"require": "./dist/index.js"
-	},
-	"directories": {
-		"lib": "src",
-		"test": "__tests__"
-	},
-	"files": [
-		"dist"
-	],
-	"contributors": [
-		"Vlad Frangu <kingdgrizzle@gmail.com>",
-		"Crawl <icrawltogo@gmail.com>",
-		"Amish Shah <amishshah.2k@gmail.com>",
-		"SpaceEEC <spaceeec@yahoo.com>",
-		"Antonio Roman <kyradiscord@gmail.com>"
-	],
-	"license": "Apache-2.0",
-	"keywords": [
-		"discord",
-		"api",
-		"bot",
-		"client",
-		"node",
-		"discordapp",
-		"discordjs"
-	],
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/discordjs/discord.js.git"
-	},
-	"bugs": {
-		"url": "https://github.com/discordjs/discord.js/issues"
-	},
-	"homepage": "https://discord.js.org",
-	"dependencies": {
-		"@sindresorhus/is": "^4.3.0",
-		"discord-api-types": "^0.26.1",
-		"ts-mixer": "^6.0.0",
-		"tslib": "^2.3.1",
-		"zod": "^3.11.6"
-	},
-	"devDependencies": {
-		"@babel/core": "^7.16.12",
-		"@babel/plugin-proposal-decorators": "^7.16.5",
-		"@babel/preset-env": "^7.16.11",
-		"@babel/preset-typescript": "^7.16.5",
-		"@discordjs/ts-docgen": "^0.3.4",
-		"@types/jest": "^27.0.3",
-		"@types/node": "^16.11.6",
-		"@typescript-eslint/eslint-plugin": "^5.10.0",
-		"@typescript-eslint/parser": "^5.10.0",
-		"babel-plugin-transform-typescript-metadata": "^0.3.2",
-		"eslint": "^8.7.0",
-		"eslint-config-marine": "^9.3.2",
-		"eslint-config-prettier": "^8.3.0",
-		"eslint-plugin-prettier": "^4.0.0",
-		"jest": "^27.4.7",
-		"prettier": "^2.5.1",
-		"standard-version": "^9.3.2",
-		"tsup": "^5.11.11",
-		"typedoc": "^0.22.11",
-		"typescript": "^4.5.5"
-	},
-	"engines": {
-		"node": ">=16.9.0"
-	},
-	"publishConfig": {
-		"access": "public"
-	}
+  "$schema": "https://json.schemastore.org/package.json",
+  "name": "@discordjs/builders",
+  "version": "1.11.2",
+  "description": "A set of builders that you can use when creating your bot",
+  "exports": {
+    ".": {
+      "require": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.js"
+      },
+      "import": {
+        "types": "./dist/index.d.mts",
+        "default": "./dist/index.mjs"
+      }
+    }
+  },
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "directories": {
+    "lib": "src",
+    "test": "__tests__"
+  },
+  "files": [
+    "dist"
+  ],
+  "contributors": [
+    "Vlad Frangu <me@vladfrangu.dev>",
+    "Crawl <icrawltogo@gmail.com>",
+    "Amish Shah <amishshah.2k@gmail.com>",
+    "SpaceEEC <spaceeec@yahoo.com>",
+    "Aura Román <kyradiscord@gmail.com>"
+  ],
+  "license": "Apache-2.0",
+  "keywords": [
+    "discord",
+    "api",
+    "bot",
+    "client",
+    "node",
+    "discordapp",
+    "discordjs"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/discordjs/discord.js.git",
+    "directory": "packages/builders"
+  },
+  "bugs": {
+    "url": "https://github.com/discordjs/discord.js/issues"
+  },
+  "homepage": "https://discord.js.org",
+  "funding": "https://github.com/discordjs/discord.js?sponsor",
+  "dependencies": {
+    "@sapphire/shapeshift": "^4.0.0",
+    "discord-api-types": "^0.38.1",
+    "fast-deep-equal": "^3.1.3",
+    "ts-mixer": "^6.0.4",
+    "tslib": "^2.6.3",
+    "@discordjs/formatters": "^0.6.1",
+    "@discordjs/util": "^1.1.1"
+  },
+  "devDependencies": {
+    "@favware/cliff-jumper": "^4.1.0",
+    "@types/node": "^16.18.105",
+    "@vitest/coverage-v8": "^2.0.5",
+    "cross-env": "^7.0.3",
+    "esbuild-plugin-version-injector": "^1.2.1",
+    "eslint": "^8.57.0",
+    "eslint-config-neon": "^0.1.62",
+    "eslint-formatter-pretty": "^6.0.1",
+    "prettier": "^3.3.3",
+    "tsup": "^8.2.4",
+    "turbo": "^2.0.14",
+    "typescript": "~5.5.4",
+    "vitest": "^2.0.5",
+    "@discordjs/api-extractor": "^7.38.1",
+    "@discordjs/scripts": "^0.1.0"
+  },
+  "engines": {
+    "node": ">=16.11.0"
+  },
+  "publishConfig": {
+    "access": "public",
+    "provenance": true
+  },
+  "scripts": {
+    "test": "vitest run",
+    "build": "tsc --noEmit && tsup",
+    "build:docs": "tsc -p tsconfig.docs.json",
+    "lint": "prettier --check . && cross-env TIMING=1 eslint --format=pretty src __tests__",
+    "format": "prettier --write . && cross-env TIMING=1 eslint --fix --format=pretty src __tests__",
+    "fmt": "pnpm run format",
+    "docs": "pnpm run build:docs && api-extractor run --local --minify && generate-split-documentation",
+    "changelog": "git cliff --prepend ./CHANGELOG.md -u -c ./cliff.toml -r ../../ --include-path 'packages/builders/*'",
+    "release": "cliff-jumper"
+  }
 }
