@@ -1,85 +1,76 @@
 {
-	"name": "@discordjs/collection",
-	"version": "0.4.0",
-	"description": "Utility data structure used in Discord.js",
-	"scripts": {
-		"pretest": "npm run build",
-		"test": "jest --pass-with-no-tests",
-		"test:ci": "jest --no-stack-trace --verbose --pass-with-no-tests",
-		"prebuild": "npm run lint",
-		"build": "tsup",
-		"lint": "eslint src --ext mjs,js,ts",
-		"lint:fix": "eslint src --ext mjs,js,ts --fix",
-		"format": "prettier --write **/*.{ts,js,json,yml,yaml}",
-		"prepare": "is-ci || husky install",
-		"docs": "typedoc --json docs/typedoc-out.json src/index.ts && node scripts/docs.mjs",
-		"prepublishOnly": "npm run lint && npm run test",
-		"release": "standard-version --preset angular"
-	},
-	"main": "./dist/index.js",
-	"module": "./dist/index.mjs",
-	"typings": "./dist/index.d.ts",
-	"exports": {
-		"import": "./dist/index.mjs",
-		"require": "./dist/index.js"
-	},
-	"directories": {
-		"lib": "src",
-		"test": "__tests__"
-	},
-	"files": [
-		"dist"
-	],
-	"contributors": [
-		"Crawl <icrawltogo@gmail.com>",
-		"Amish Shah <amishshah.2k@gmail.com>",
-		"SpaceEEC <spaceeec@yahoo.com>",
-		"Vlad Frangu <kingdgrizzle@gmail.com>"
-	],
-	"license": "Apache-2.0",
-	"keywords": [
-		"map",
-		"collection",
-		"utility"
-	],
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/discordjs/collection.git"
-	},
-	"bugs": {
-		"url": "https://github.com/discordjs/collection/issues"
-	},
-	"homepage": "https://github.com/discordjs/collection",
-	"devDependencies": {
-		"@babel/core": "^7.16.5",
-		"@babel/preset-env": "^7.16.5",
-		"@babel/preset-typescript": "^7.16.5",
-		"@commitlint/cli": "^15.0.0",
-		"@commitlint/config-angular": "^15.0.0",
-		"@discordjs/ts-docgen": "^0.3.4",
-		"@types/jest": "^27.0.3",
-		"@types/node": "^16.11.6",
-		"@typescript-eslint/eslint-plugin": "^5.8.0",
-		"@typescript-eslint/parser": "^5.8.0",
-		"eslint": "^8.5.0",
-		"eslint-config-marine": "^9.1.0",
-		"eslint-config-prettier": "^8.3.0",
-		"eslint-plugin-prettier": "^4.0.0",
-		"husky": "^7.0.4",
-		"is-ci": "^3.0.1",
-		"jest": "^27.4.5",
-		"lint-staged": "^12.1.4",
-		"prettier": "^2.5.1",
-		"standard-version": "^9.3.2",
-		"tsup": "^5.11.8",
-		"typedoc": "^0.22.10",
-		"typescript": "^4.5.4"
-	},
-	"engines": {
-		"node": ">=16.0.0",
-		"npm": ">=7.0.0"
-	},
-	"publishConfig": {
-		"access": "public"
-	}
+  "name": "@discordjs/collection",
+  "version": "1.5.3",
+  "description": "Utility data structure used in discord.js",
+  "scripts": {
+    "test": "vitest run",
+    "build": "tsup",
+    "build:docs": "tsc -p tsconfig.docs.json",
+    "lint": "prettier --check . && cross-env TIMING=1 eslint src __tests__ --ext .mjs,.js,.ts --format=pretty",
+    "format": "prettier --write . && cross-env TIMING=1 eslint src __tests__ --ext .mjs,.js,.ts --fix --format=pretty",
+    "fmt": "yarn format",
+    "docs": "yarn build:docs && api-extractor run --local && api-extractor run --local --config ./api-extractor-docs.json",
+    "prepack": "yarn lint && yarn test && yarn build",
+    "changelog": "git cliff --prepend ./CHANGELOG.md -u -c ./cliff.toml -r ../../ --include-path 'packages/collection/*'",
+    "release": "cliff-jumper"
+  },
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    "types": "./dist/index.d.ts",
+    "import": "./dist/index.mjs",
+    "require": "./dist/index.js"
+  },
+  "directories": {
+    "lib": "src",
+    "test": "__tests__"
+  },
+  "files": [
+    "dist"
+  ],
+  "contributors": [
+    "Crawl <icrawltogo@gmail.com>",
+    "Amish Shah <amishshah.2k@gmail.com>",
+    "SpaceEEC <spaceeec@yahoo.com>",
+    "Vlad Frangu <kingdgrizzle@gmail.com>",
+    "Aura Román <kyradiscord@gmail.com>"
+  ],
+  "license": "Apache-2.0",
+  "keywords": [
+    "map",
+    "collection",
+    "utility"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/discordjs/discord.js.git",
+    "directory": "packages/collection"
+  },
+  "bugs": {
+    "url": "https://github.com/discordjs/discord.js/issues"
+  },
+  "homepage": "https://discord.js.org",
+  "devDependencies": {
+    "@favware/cliff-jumper": "^2.1.1",
+    "@microsoft/api-extractor": "^7.36.4",
+    "@types/node": "16.18.40",
+    "@vitest/coverage-v8": "^0.34.2",
+    "cross-env": "^7.0.3",
+    "esbuild-plugin-version-injector": "^1.2.0",
+    "eslint": "^8.47.0",
+    "eslint-config-neon": "^0.1.47",
+    "eslint-formatter-pretty": "^5.0.0",
+    "prettier": "^2.8.8",
+    "tsup": "^7.2.0",
+    "turbo": "^1.10.12",
+    "typescript": "^5.1.6",
+    "vitest": "^0.34.2"
+  },
+  "engines": {
+    "node": ">=16.11.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
 }
