@@ -1,89 +1,119 @@
 {
-	"name": "@discordjs/rest",
-	"version": "0.3.0",
-	"description": "The REST API for discord.js",
-	"scripts": {
-		"build": "tsup && tsc --emitDeclarationOnly --incremental",
-		"test": "jest --pass-with-no-tests --collect-coverage",
-		"lint": "prettier --check . && eslint src __tests__ --ext mjs,js,ts",
-		"format": "prettier --write . && eslint src __tests__ --ext mjs,js,ts --fix",
-		"docs": "typedoc --json docs/typedoc-out.json src/index.ts && node scripts/docs.mjs",
-		"prepublishOnly": "yarn build && yarn lint && yarn test",
-		"changelog": "git cliff --prepend ./CHANGELOG.md -u -c ./cliff.toml -r ../../ --include-path 'packages/rest/*'"
-	},
-	"main": "./dist/index.js",
-	"module": "./dist/index.mjs",
-	"typings": "./dist/index.d.ts",
-	"exports": {
-		"import": "./dist/index.mjs",
-		"require": "./dist/index.js"
-	},
-	"directories": {
-		"lib": "src",
-		"test": "__tests__"
-	},
-	"files": [
-		"dist"
-	],
-	"contributors": [
-		"Crawl <icrawltogo@gmail.com>",
-		"Amish Shah <amishshah.2k@gmail.com>",
-		"SpaceEEC <spaceeec@yahoo.com>",
-		"Vlad Frangu <kingdgrizzle@gmail.com>",
-		"Antonio Roman <kyradiscord@gmail.com>"
-	],
-	"license": "Apache-2.0",
-	"keywords": [
-		"discord",
-		"api",
-		"rest",
-		"discordapp",
-		"discordjs"
-	],
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/discordjs/discord.js.git"
-	},
-	"bugs": {
-		"url": "https://github.com/discordjs/discord.js/issues"
-	},
-	"homepage": "https://discord.js.org",
-	"dependencies": {
-		"@discordjs/collection": "^0.4.0",
-		"@sapphire/async-queue": "^1.1.9",
-		"@sapphire/snowflake": "^3.0.1",
-		"discord-api-types": "^0.26.1",
-		"form-data": "^4.0.0",
-		"node-fetch": "^2.6.5",
-		"tslib": "^2.3.1"
-	},
-	"devDependencies": {
-		"@babel/core": "^7.16.12",
-		"@babel/plugin-proposal-decorators": "^7.16.7",
-		"@babel/preset-env": "^7.16.11",
-		"@babel/preset-typescript": "^7.16.7",
-		"@discordjs/ts-docgen": "^0.3.4",
-		"@types/jest": "^27.4.0",
-		"@types/node-fetch": "^2.5.10",
-		"@typescript-eslint/eslint-plugin": "^5.10.0",
-		"@typescript-eslint/parser": "^5.10.0",
-		"babel-plugin-const-enum": "^1.2.0",
-		"babel-plugin-transform-typescript-metadata": "^0.3.2",
-		"eslint": "^8.7.0",
-		"eslint-config-marine": "^9.3.2",
-		"eslint-config-prettier": "^8.3.0",
-		"eslint-plugin-prettier": "^4.0.0",
-		"jest": "^27.4.7",
-		"nock": "^13.2.2",
-		"prettier": "^2.5.1",
-		"tsup": "^5.11.11",
-		"typedoc": "^0.22.11",
-		"typescript": "^4.5.5"
-	},
-	"engines": {
-		"node": ">=16.9.0"
-	},
-	"publishConfig": {
-		"access": "public"
-	}
+  "$schema": "https://json.schemastore.org/package.json",
+  "name": "@discordjs/rest",
+  "version": "2.5.1",
+  "description": "The REST API for discord.js",
+  "exports": {
+    ".": {
+      "node": {
+        "require": {
+          "types": "./dist/index.d.ts",
+          "default": "./dist/index.js"
+        },
+        "import": {
+          "types": "./dist/index.d.mts",
+          "default": "./dist/index.mjs"
+        }
+      },
+      "default": {
+        "require": {
+          "types": "./dist/web.d.ts",
+          "default": "./dist/web.js"
+        },
+        "import": {
+          "types": "./dist/web.d.mts",
+          "default": "./dist/web.mjs"
+        }
+      }
+    },
+    "./*": {
+      "require": {
+        "types": "./dist/strategies/*.d.ts",
+        "default": "./dist/strategies/*.js"
+      },
+      "import": {
+        "types": "./dist/strategies/*.d.mts",
+        "default": "./dist/strategies/*.mjs"
+      }
+    }
+  },
+  "types": "./dist/index.d.ts",
+  "directories": {
+    "lib": "src",
+    "test": "__tests__"
+  },
+  "files": [
+    "dist"
+  ],
+  "contributors": [
+    "Crawl <icrawltogo@gmail.com>",
+    "Amish Shah <amishshah.2k@gmail.com>",
+    "SpaceEEC <spaceeec@yahoo.com>",
+    "Vlad Frangu <me@vladfrangu.dev>",
+    "Aura Román <kyradiscord@gmail.com>"
+  ],
+  "license": "Apache-2.0",
+  "keywords": [
+    "discord",
+    "api",
+    "rest",
+    "discordapp",
+    "discordjs"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/discordjs/discord.js.git",
+    "directory": "packages/rest"
+  },
+  "bugs": {
+    "url": "https://github.com/discordjs/discord.js/issues"
+  },
+  "homepage": "https://discord.js.org",
+  "funding": "https://github.com/discordjs/discord.js?sponsor",
+  "dependencies": {
+    "@sapphire/async-queue": "^1.5.3",
+    "@sapphire/snowflake": "^3.5.3",
+    "@vladfrangu/async_event_emitter": "^2.4.6",
+    "discord-api-types": "^0.38.1",
+    "magic-bytes.js": "^1.10.0",
+    "tslib": "^2.6.3",
+    "undici": "6.21.3",
+    "@discordjs/collection": "^2.1.1",
+    "@discordjs/util": "^1.1.1"
+  },
+  "devDependencies": {
+    "@favware/cliff-jumper": "^4.1.0",
+    "@types/node": "18.17.9",
+    "@vitest/coverage-v8": "^2.0.5",
+    "cross-env": "^7.0.3",
+    "esbuild-plugin-version-injector": "^1.2.1",
+    "eslint": "^8.57.0",
+    "eslint-config-neon": "^0.1.62",
+    "eslint-formatter-pretty": "^6.0.1",
+    "prettier": "^3.3.3",
+    "tsup": "^8.2.4",
+    "turbo": "^2.0.14",
+    "typescript": "~5.5.4",
+    "vitest": "^2.0.5",
+    "@discordjs/scripts": "^0.1.0",
+    "@discordjs/api-extractor": "^7.38.1"
+  },
+  "engines": {
+    "node": ">=18"
+  },
+  "publishConfig": {
+    "access": "public",
+    "provenance": true
+  },
+  "scripts": {
+    "test": "vitest run",
+    "build": "tsc --noEmit && tsup",
+    "build:docs": "tsc -p tsconfig.docs.json",
+    "lint": "prettier --check . && cross-env TIMING=1 eslint --format=pretty src __tests__",
+    "format": "prettier --write . && cross-env TIMING=1 eslint --fix --format=pretty src __tests__",
+    "fmt": "pnpm run format",
+    "docs": "pnpm run build:docs && api-extractor run --local --minify && generate-split-documentation",
+    "changelog": "git cliff --prepend ./CHANGELOG.md -u -c ./cliff.toml -r ../../ --include-path 'packages/rest/*'",
+    "release": "cliff-jumper"
+  }
 }
