@@ -44,47 +44,64 @@
   ],
   "repository": {
     "type": "git",
-    "url": "https://github.com/discordjs/discord.js.git"
+    "url": "https://github.com/discordjs/discord.js.git",
+    "directory": "packages/discord.js"
   },
   "bugs": {
     "url": "https://github.com/discordjs/discord.js/issues"
   },
   "homepage": "https://discord.js.org",
+  "funding": "https://github.com/discordjs/discord.js?sponsor",
   "dependencies": {
-    "@discordjs/builders": "^0.16.0",
-    "@discordjs/collection": "^0.7.0",
-    "@sapphire/async-queue": "^1.5.0",
-    "@types/node-fetch": "^2.6.3",
-    "@types/ws": "^8.5.4",
-    "discord-api-types": "^0.33.5",
-    "form-data": "^4.0.0",
-    "node-fetch": "^2.6.7",
-    "ws": "^8.13.0"
+    "@discordjs/builders": "^1.11.2",
+    "@discordjs/collection": "1.5.3",
+    "@discordjs/formatters": "^0.6.1",
+    "@discordjs/ws": "^1.2.3",
+    "@sapphire/snowflake": "3.5.3",
+    "discord-api-types": "^0.38.1",
+    "fast-deep-equal": "3.1.3",
+    "lodash.snakecase": "4.1.1",
+    "magic-bytes.js": "^1.10.0",
+    "tslib": "^2.6.3",
+    "undici": "6.21.3",
+    "@discordjs/rest": "^2.5.1",
+    "@discordjs/util": "^1.1.1"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.5.1",
-    "@commitlint/config-angular": "^17.4.4",
-    "@discordjs/docgen": "^0.11.1",
-    "@favware/cliff-jumper": "^2.0.0",
-    "@favware/npm-deprecate": "^1.0.7",
-    "@types/node": "^16.11.45",
-    "conventional-changelog-cli": "^2.2.2",
-    "dtslint": "^4.2.1",
-    "eslint": "^8.37.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-prettier": "^4.2.1",
-    "husky": "^8.0.3",
-    "is-ci": "^3.0.1",
-    "jest": "^29.5.0",
-    "lint-staged": "^13.2.0",
-    "prettier": "^2.8.7",
-    "tsd": "^0.28.1",
-    "tslint": "^6.1.3",
-    "typescript": "^5.0.3"
+    "@favware/cliff-jumper": "^4.1.0",
+    "@types/node": "^16.18.105",
+    "@typescript-eslint/eslint-plugin": "^8.2.0",
+    "@typescript-eslint/parser": "^8.2.0",
+    "cross-env": "^7.0.3",
+    "dtslint": "4.2.1",
+    "eslint": "^8.57.0",
+    "eslint-formatter-pretty": "^5.0.0",
+    "jest": "29.7.0",
+    "prettier": "^3.3.3",
+    "tsd": "^0.31.1",
+    "tslint": "6.1.3",
+    "turbo": "^2.0.14",
+    "typescript": "~5.5.4",
+    "@discordjs/api-extractor": "^7.38.1",
+    "@discordjs/docgen": "^0.12.1",
+    "@discordjs/scripts": "^0.1.0"
   },
   "engines": {
-    "node": ">=16.6.0",
-    "npm": ">=7.0.0"
+    "node": ">=18"
+  },
+  "publishConfig": {
+    "provenance": true
+  },
+  "scripts": {
+    "test": "pnpm run docs:test && pnpm run test:typescript",
+    "test:typescript": "tsc --noEmit && tsd",
+    "lint": "prettier --check . && tslint typings/index.d.ts && cross-env ESLINT_USE_FLAT_CONFIG=false eslint --format=pretty src typings",
+    "format": "prettier --write . && cross-env ESLINT_USE_FLAT_CONFIG=false eslint --fix --format=pretty src",
+    "fmt": "pnpm run format",
+    "docs": "docgen -i \"./src/*.js\" \"./src/**/*.js\" -c ./docs/index.json -r ../../ -o ./docs/docs.json && pnpm run docs:new",
+    "docs:test": "docgen -i \"./src/*.js\" \"./src/**/*.js\" -c ./docs/index.json -r ../../",
+    "docs:new": "api-extractor run --local --minify && generate-split-documentation",
+    "changelog": "git cliff --prepend ./CHANGELOG.md -u -c ./cliff.toml -r ../../ --include-path 'packages/discord.js/*'",
+    "release": "cliff-jumper"
   }
 }
