@@ -1,25 +1,29 @@
 {
   "name": "dotenv",
-  "version": "16.0.0",
+  "version": "16.6.1",
   "description": "Loads environment variables from .env file",
   "main": "lib/main.js",
   "types": "lib/main.d.ts",
   "exports": {
     ".": {
-      "require": "./lib/main.js",
       "types": "./lib/main.d.ts",
+      "require": "./lib/main.js",
       "default": "./lib/main.js"
     },
     "./config": "./config.js",
     "./config.js": "./config.js",
+    "./lib/env-options": "./lib/env-options.js",
+    "./lib/env-options.js": "./lib/env-options.js",
+    "./lib/cli-options": "./lib/cli-options.js",
+    "./lib/cli-options.js": "./lib/cli-options.js",
     "./package.json": "./package.json"
   },
   "scripts": {
     "dts-check": "tsc --project tests/types/tsconfig.json",
     "lint": "standard",
-    "lint-readme": "standard-markdown",
     "pretest": "npm run lint && npm run dts-check",
-    "test": "tap tests/*.js --100 -Rspec",
+    "test": "tap run --allow-empty-coverage --disable-coverage --timeout=60000",
+    "test:coverage": "tap run --show-full-coverage --timeout=60000 --coverage-report=text --coverage-report=lcov",
     "prerelease": "npm test",
     "release": "standard-version"
   },
@@ -27,6 +31,8 @@
     "type": "git",
     "url": "git://github.com/motdotla/dotenv.git"
   },
+  "homepage": "https://github.com/motdotla/dotenv#readme",
+  "funding": "https://dotenvx.com",
   "keywords": [
     "dotenv",
     "env",
@@ -39,17 +45,18 @@
   "readmeFilename": "README.md",
   "license": "BSD-2-Clause",
   "devDependencies": {
-    "@types/node": "^17.0.9",
-    "decache": "^4.6.1",
-    "dtslint": "^3.7.0",
-    "sinon": "^12.0.1",
-    "standard": "^16.0.4",
-    "standard-markdown": "^7.1.0",
-    "standard-version": "^9.3.2",
-    "tap": "^15.1.6",
-    "typescript": "^4.5.4"
+    "@types/node": "^18.11.3",
+    "decache": "^4.6.2",
+    "sinon": "^14.0.1",
+    "standard": "^17.0.0",
+    "standard-version": "^9.5.0",
+    "tap": "^19.2.0",
+    "typescript": "^4.8.4"
   },
   "engines": {
     "node": ">=12"
+  },
+  "browser": {
+    "fs": false
   }
 }
