@@ -1,89 +1,131 @@
 {
-    "name": "node-fetch",
-    "version": "2.7.0",
-    "description": "A light-weight module that brings window.fetch to node.js",
-    "main": "lib/index.js",
-    "browser": "./browser.js",
-    "module": "lib/index.mjs",
-    "files": [
-        "lib/index.js",
-        "lib/index.mjs",
-        "lib/index.es.js",
-        "browser.js"
-    ],
-    "engines": {
-        "node": "4.x || >=6.0.0"
-    },
-    "scripts": {
-        "build": "cross-env BABEL_ENV=rollup rollup -c",
-        "prepare": "npm run build",
-        "test": "cross-env BABEL_ENV=test mocha --require babel-register --throw-deprecation test/test.js",
-        "report": "cross-env BABEL_ENV=coverage nyc --reporter lcov --reporter text mocha -R spec test/test.js",
-        "coverage": "cross-env BABEL_ENV=coverage nyc --reporter json --reporter text mocha -R spec test/test.js && codecov -f coverage/coverage-final.json"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/bitinn/node-fetch.git"
-    },
-    "keywords": [
-        "fetch",
-        "http",
-        "promise"
-    ],
-    "author": "David Frank",
-    "license": "MIT",
-    "bugs": {
-        "url": "https://github.com/bitinn/node-fetch/issues"
-    },
-    "homepage": "https://github.com/bitinn/node-fetch",
-    "dependencies": {
-        "whatwg-url": "^5.0.0"
-    },
-    "peerDependencies": {
-        "encoding": "^0.1.0"
-    },
-    "peerDependenciesMeta": {
-        "encoding": {
-            "optional": true
-        }
-    },
-    "devDependencies": {
-        "@ungap/url-search-params": "^0.1.2",
-        "abort-controller": "^1.1.0",
-        "abortcontroller-polyfill": "^1.3.0",
-        "babel-core": "^6.26.3",
-        "babel-plugin-istanbul": "^4.1.6",
-        "babel-plugin-transform-async-generator-functions": "^6.24.1",
-        "babel-polyfill": "^6.26.0",
-        "babel-preset-env": "1.4.0",
-        "babel-register": "^6.16.3",
-        "chai": "^3.5.0",
-        "chai-as-promised": "^7.1.1",
-        "chai-iterator": "^1.1.1",
-        "chai-string": "~1.3.0",
-        "codecov": "3.3.0",
-        "cross-env": "^5.2.0",
-        "form-data": "^2.3.3",
-        "is-builtin-module": "^1.0.0",
-        "mocha": "^5.0.0",
-        "nyc": "11.9.0",
-        "parted": "^0.1.1",
-        "promise": "^8.0.3",
-        "resumer": "0.0.0",
-        "rollup": "^0.63.4",
-        "rollup-plugin-babel": "^3.0.7",
-        "string-to-arraybuffer": "^1.0.2",
-        "teeny-request": "3.7.0"
-    },
-    "release": {
-        "branches": [
-            "+([0-9]).x",
-            "main",
-            "next",
-            {
-                "name": "beta",
-                "prerelease": true
-            }
-        ]
+  "name": "node-fetch",
+  "version": "3.3.2",
+  "description": "A light-weight module that brings Fetch API to node.js",
+  "main": "./src/index.js",
+  "sideEffects": false,
+  "type": "module",
+  "files": [
+    "src",
+    "@types/index.d.ts"
+  ],
+  "types": "./@types/index.d.ts",
+  "engines": {
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
+  },
+  "scripts": {
+    "test": "mocha",
+    "coverage": "c8 report --reporter=text-lcov | coveralls",
+    "test-types": "tsd",
+    "lint": "xo"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/node-fetch/node-fetch.git"
+  },
+  "keywords": [
+    "fetch",
+    "http",
+    "promise",
+    "request",
+    "curl",
+    "wget",
+    "xhr",
+    "whatwg"
+  ],
+  "author": "David Frank",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/node-fetch/node-fetch/issues"
+  },
+  "homepage": "https://github.com/node-fetch/node-fetch",
+  "funding": {
+    "type": "opencollective",
+    "url": "https://opencollective.com/node-fetch"
+  },
+  "devDependencies": {
+    "abort-controller": "^3.0.0",
+    "abortcontroller-polyfill": "^1.7.1",
+    "busboy": "^1.4.0",
+    "c8": "^7.7.2",
+    "chai": "^4.3.4",
+    "chai-as-promised": "^7.1.1",
+    "chai-iterator": "^3.0.2",
+    "chai-string": "^1.5.0",
+    "coveralls": "^3.1.0",
+    "form-data": "^4.0.0",
+    "formdata-node": "^4.2.4",
+    "mocha": "^9.1.3",
+    "p-timeout": "^5.0.0",
+    "stream-consumers": "^1.0.1",
+    "tsd": "^0.14.0",
+    "xo": "^0.39.1"
+  },
+  "dependencies": {
+    "data-uri-to-buffer": "^4.0.0",
+    "fetch-blob": "^3.1.4",
+    "formdata-polyfill": "^4.0.10"
+  },
+  "tsd": {
+    "cwd": "@types",
+    "compilerOptions": {
+      "esModuleInterop": true
     }
+  },
+  "xo": {
+    "envs": [
+      "node",
+      "browser"
+    ],
+    "ignores": [
+      "example.js"
+    ],
+    "rules": {
+      "complexity": 0,
+      "import/extensions": 0,
+      "import/no-useless-path-segments": 0,
+      "import/no-anonymous-default-export": 0,
+      "import/no-named-as-default": 0,
+      "unicorn/import-index": 0,
+      "unicorn/no-array-reduce": 0,
+      "unicorn/prefer-node-protocol": 0,
+      "unicorn/numeric-separators-style": 0,
+      "unicorn/explicit-length-check": 0,
+      "capitalized-comments": 0,
+      "node/no-unsupported-features/es-syntax": 0,
+      "@typescript-eslint/member-ordering": 0
+    },
+    "overrides": [
+      {
+        "files": "test/**/*.js",
+        "envs": [
+          "node",
+          "mocha"
+        ],
+        "rules": {
+          "max-nested-callbacks": 0,
+          "no-unused-expressions": 0,
+          "no-warning-comments": 0,
+          "new-cap": 0,
+          "guard-for-in": 0,
+          "unicorn/no-array-for-each": 0,
+          "unicorn/prevent-abbreviations": 0,
+          "promise/prefer-await-to-then": 0,
+          "ava/no-import-test-files": 0
+        }
+      }
+    ]
+  },
+  "runkitExampleFilename": "example.js",
+  "release": {
+    "branches": [
+      "+([0-9]).x",
+      "main",
+      "next",
+      {
+        "name": "beta",
+        "prerelease": true
+      }
+    ]
+  }
 }
